Switch to 8-byte key/8-byte value for maps

+++ examples/hashmap/include/atomicmap.h
@@ -15,5 +15,5 @@ public:
         bool get_interesting_colors(std::vector<workload_config>* workload, std::vector<ColorID>& interesting_colors);
         void init_synchronizer(std::vector<std::string>* log_addr, std::vector<ColorID>& interesting_colors, bool replication);
 
-        uint32_t get(uint32_t key);
+        uint64_t get(uint64_t key);
 };
