+++ examples/hashmap/src/atomicmap_workload.cc
@@ -5,7 +5,7 @@
 #include <atomicmap.h>
 
 void ycsb_insert::Run() {
-        uint32_t key, value;
+        uint64_t key, value;
         assert(m_map != NULL);
        
         // key = (start, end)
@@ -18,7 +18,7 @@ void ycsb_insert::Run() {
 }
 
 write_id ycsb_insert::AsyncRun() {
-        uint32_t key, value;
+        uint64_t key, value;
         assert(m_map != NULL);
         assert(m_dep_color == NULL);            // XXX: in AtomicMap, do not support cross-color put at the moment
        
@@ -42,7 +42,7 @@ write_id ycsb_insert::AsyncWeaklyConsistentRun() {
 }
 
 void ycsb_read::Run() {
-        uint32_t key;
+        uint64_t key;
         assert(m_map != NULL);
        
         // key = (start, end)
@@ -52,7 +52,7 @@ void ycsb_read::Run() {
 }
 
 write_id ycsb_read::AsyncRun() {
-        uint32_t key;
+        uint64_t key;
         assert(m_map != NULL);
        
         // key = (start, end)
@@ -75,8 +75,8 @@ write_id ycsb_read::AsyncWeaklyConsistentRun() {
 }
 
 Txn** atomicmap_workload_generator::Gen() {
-        uint32_t i;
-        uint32_t total_op_count;
+        uint64_t i;
+        uint64_t total_op_count;
         double r;
 
         // percent of single operations
@@ -88,8 +88,8 @@ Txn** atomicmap_workload_generator::Gen() {
 
         // make transactions
         vector<double> proportions;
-        vector<uint32_t> allocations;
-        uint32_t n = 0;
+        vector<uint64_t> allocations;
+        uint64_t n = 0;
         for (auto w : *m_workload) {
                 n += w.op_count; 
                 double p = static_cast<double>(n) / total_op_count;
@@ -106,8 +106,8 @@ Txn** atomicmap_workload_generator::Gen() {
                         if (proportions[j] > r && allocations[j] > 0) { 
                                 std::string op_type = (*m_workload)[j].op_type; 
                                 struct colors* color = &(*m_workload)[j].first_color; 
-                                uint32_t start = 0; 
-                                uint32_t end = m_range; 
+                                uint64_t start = 0; 
+                                uint64_t end = m_range; 
 
                                 if (op_type == "get") {
                                         txns[i] = new ycsb_read(m_map, color, start, end, m_context, Txn::optype::GET);
